docker/google-auth/requirements.txt
Flask
flask-session
google_api_python_client==2.66.0
google_auth_oauthlib==0.5.2
protobuf==3.20.3
redis==5.0.1
Requests==2.31.0
